Cleanupd up some compiler warnings

+++ src/display/display.cpp
@@ -169,16 +169,16 @@ Display::fill_rounded_rect(const Rectf& rect, float radius, const Color& color)
   glBegin(GL_QUAD_STRIP);
   for(int i = 0; i <= n; ++i)
     {
-      float x = sinf(i * (math::pi/2) / n) * radius;
-      float y = cosf(i * (math::pi/2) / n) * radius;
+      float x = sinf(static_cast<float>(i) * (math::pi/2.0f) / static_cast<float>(n)) * radius;
+      float y = cosf(static_cast<float>(i) * (math::pi/2.0f) / static_cast<float>(n)) * radius;
 
       glVertex2f(irect.left  - x, irect.top - y);
       glVertex2f(irect.right + x, irect.top - y);
     }
   for(int i = 0; i <= n; ++i)
     {
-      float x = cosf(i * (math::pi/2) / n) * radius;
-      float y = sinf(i * (math::pi/2) / n) * radius;
+      float x = cosf(static_cast<float>(i) * (math::pi/2.0f) / static_cast<float>(n)) * radius;
+      float y = sinf(static_cast<float>(i) * (math::pi/2.0f) / static_cast<float>(n)) * radius;
 
       glVertex2f(irect.left  - x, irect.bottom + y);
       glVertex2f(irect.right + x, irect.bottom + y);
@@ -211,31 +211,31 @@ Display::draw_rounded_rect(const Rectf& rect, float radius, const Color& color)
   glBegin(GL_LINE_STRIP);
   for(int i = 0; i <= n; ++i)
     {
-      float x = sinf(i * (math::pi/2) / n) * radius;
-      float y = cosf(i * (math::pi/2) / n) * radius;
+      float x = sinf(static_cast<float>(i) * (math::pi/2) / static_cast<float>(n)) * radius;
+      float y = cosf(static_cast<float>(i) * (math::pi/2) / static_cast<float>(n)) * radius;
 
       glVertex2f(irect.left  - x, irect.top - y);
     }
   for(int i = 0; i <= n; ++i)
     {
-      float x = cosf(i * (math::pi/2) / n) * radius;
-      float y = sinf(i * (math::pi/2) / n) * radius;
+      float x = cosf(static_cast<float>(i) * (math::pi/2) / static_cast<float>(n)) * radius;
+      float y = sinf(static_cast<float>(i) * (math::pi/2) / static_cast<float>(n)) * radius;
 
       glVertex2f(irect.left  - x, irect.bottom + y);
     }
 
   for(int i = 0; i <= n; ++i)
     {
-      float x = sinf(i * (math::pi/2) / n) * radius;
-      float y = cosf(i * (math::pi/2) / n) * radius;
+      float x = sinf(static_cast<float>(i) * (math::pi/2) / static_cast<float>(n)) * radius;
+      float y = cosf(static_cast<float>(i) * (math::pi/2) / static_cast<float>(n)) * radius;
 
       glVertex2f(irect.right + x, irect.bottom + y);
     }
 
   for(int i = 0; i <= n; ++i)
     {
-      float x = cosf(i * (math::pi/2) / n) * radius;
-      float y = sinf(i * (math::pi/2) / n) * radius;
+      float x = cosf(static_cast<float>(i) * (math::pi/2) / static_cast<float>(n)) * radius;
+      float y = sinf(static_cast<float>(i) * (math::pi/2) / static_cast<float>(n)) * radius;
         
       glVertex2f(irect.right + x, irect.top - y);
     }
@@ -270,13 +270,13 @@ Display::draw_circle(const Vector2f& pos, float radius, const Color& color, int 
   state.color(color);
   state.activate();
 
-  float n = segments/4.0f;
+  float n = static_cast<float>(segments) / 4.0f;
   glBegin(GL_LINE_STRIP);
   glVertex2f(radius + pos.x, pos.y);
   for(int i = 1; i < segments; ++i)
     {
-      float x = cosf(i * (math::pi/2) / n) * radius;
-      float y = sinf(i * (math::pi/2) / n) * radius;
+      float x = cosf(static_cast<float>(i) * (math::pi/2) / static_cast<float>(n)) * radius;
+      float y = sinf(static_cast<float>(i) * (math::pi/2) / static_cast<float>(n)) * radius;
       
       glVertex2f(x + pos.x, y + pos.y);
     }
